+++ pandas/indexes/multi.py
@@ -19,12 +19,10 @@ from pandas.types.common import (_ensure_int64,
                                  is_list_like,
                                  is_scalar)
 from pandas.types.missing import isnull, array_equivalent
+from pandas.errors import PerformanceWarning, UnsortedIndexError
 from pandas.core.common import (_values_from_object,
                                 is_bool_indexer,
-                                is_null_slice,
-                                PerformanceWarning,
-                                UnsortedIndexError)
-
+                                is_null_slice)
 
 import pandas.core.base as base
 from pandas.util.decorators import (Appender, cache_readonly,
