+++ pandas/io/html.py
@@ -13,7 +13,8 @@ from distutils.version import LooseVersion
 import numpy as np
 
 from pandas.types.common import is_list_like
-from pandas.io.common import (EmptyDataError, _is_url, urlopen,
+from pandas.errors import EmptyDataError
+from pandas.io.common import (_is_url, urlopen,
                               parse_url, _validate_header_arg)
 from pandas.io.parsers import TextParser
 from pandas.compat import (lrange, lmap, u, string_types, iteritems,
