+++ pandas/tests/io/parser/skiprows.py
@@ -12,7 +12,7 @@ import numpy as np
 import pandas.util.testing as tm
 
 from pandas import DataFrame
-from pandas.io.common import EmptyDataError
+from pandas.errors import EmptyDataError
 from pandas.compat import StringIO, range, lrange
