+++ pandas/tests/io/test_common.py
@@ -11,7 +11,6 @@ from pandas.io import common
 from pandas.compat import is_platform_windows, StringIO
 
 from pandas import read_csv, concat
-import pandas as pd
 
 try:
     from pathlib import Path
@@ -88,23 +87,6 @@ bar2,12,13,14,15
         first = next(it)
         tm.assert_frame_equal(first, expected.iloc[[0]])
         tm.assert_frame_equal(concat(it), expected.iloc[1:])
-
-    def test_error_rename(self):
-        # see gh-12665
-        try:
-            raise common.CParserError()
-        except common.ParserError:
-            pass
-
-        try:
-            raise common.ParserError()
-        except common.CParserError:
-            pass
-
-        try:
-            raise common.ParserError()
-        except pd.parser.CParserError:
-            pass
 
 
 class TestMMapWrapper(tm.TestCase):
