+++ pandas/tslib.py
@@ -3,6 +3,6 @@
 import warnings
 warnings.warn("The pandas.tslib module is deprecated and will be "
               "removed in a future version. Please import from "
-              "the pandas._libs.tslib instead", FutureWarning, stacklevel=2)
+              "the pandas or pandas.api.exceptions instead", FutureWarning, stacklevel=2)
 from pandas._libs.tslib import (Timestamp, Timedelta,
                                NaT, OutOfBoundsDatetime)
