+++ setup.py
@@ -637,6 +637,7 @@ setup(name=DISTNAME,
                 'pandas.computation',
                 'pandas.core',
                 'pandas.indexes',
+                'pandas.errors',
                 'pandas.io',
                 'pandas.io.json',
                 'pandas.io.sas',
